move to_fp6 to dtypes/
